refactor: move playground outside packages
