+++ package.json
@@ -12,25 +12,27 @@
   ],
   "scripts": {
     "abell": "pnpm --filter abell dev",
-    "abell:build": "pnpm --filter ...abell clear-cache && pnpm --filter abell build",
+    "abell:build": "pnpm --filter abell build",
     "docs:dev": "pnpm abell:build && pnpm --filter docs dev",
     "docs:generate": "pnpm --filter docs generate",
     "// Linting Stuff Starts from Here": "",
     "eslint": "eslint packages/**/src/**/**/ --ext ts,js",
     "prettier": "prettier --write \"packages/**/**/**/*.[jt]s\"",
-    "test:once": "pnpm abell test:once && pnpm play test:once && pnpm create-abell test:once",
+    "test:playground": "cd playground && vitest",
+    "test:playground:once": "cd playground && vitest run",
+    "test:once": "pnpm test:playground:once",
     "format": "npm run prettier",
     "prepublishOnly": "echo \"NOT FROM THE ROOT!! sorry for caps hehe\" && fail"
   },
-  "dependencies": {},
   "devDependencies": {
+    "@types/jsdom": "^20.0.0",
     "@typescript-eslint/eslint-plugin": "^4.25.0",
     "@typescript-eslint/parser": "^4.25.0",
     "eslint": "^7.0.0",
     "eslint-config-google": "^0.14.0",
     "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-prettier": "^3.1.3",
-    "husky": "^4.2.5",
+    "jsdom": "^20.0.1",
     "prettier": "^2.0.5",
     "typescript": "^4.3.2",
     "vitest": "^0.3.6"
