--- packages/playground/basic/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "basic",
-  "version": "1.0.0",
-  "description": "",
-  "type": "module",
-  "scripts": {
-    "dev": "npx abell --version && npx abell dev",
-    "generate": "npx abell generate"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC"
-}
