--- packages/playground/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "playground",
-  "version": "0.0.0",
-  "description": "",
-  "scripts": {
-    "test": "vitest",
-    "test:once": "vitest run",
-    "dev:abell": "yarn workspace abell dev",
-    "build:abell": "yarn workspace abell build",
-    "generate:basic": "cd basic && yarn generate"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "abell": "workspace:*"
-  },
-  "devDependencies": {
-    "@types/jsdom": "^16.2.14",
-    "jsdom": "^19.0.0"
-  }
-}
