+++ playground/basic/_navbar.abell
@@ -7,6 +7,6 @@
 
 <style>
 ul {
-  background-color: #ddd;
+  background-color: #eee;
 }
 </style>
