+++ playground/basic/test-utils.ts
@@ -7,7 +7,7 @@ const isWindows = /^win/.test(process.platform);
 
 const windowsifyCommand = (command: string): string => {
   if (isWindows) {
-    return command.replace('yarn', 'yarn.cmd');
+    return command.replace('pnpm', 'pnpm.cmd');
   }
 
   return command;
@@ -17,7 +17,7 @@ let dir: string = __dirname;
 export async function run(cwd: string): Promise<void> {
   dir = cwd;
   return new Promise((resolve, reject) => {
-    const child = spawn(windowsifyCommand('yarn'), ['generate'], {
+    const child = spawn(windowsifyCommand('pnpm'), ['generate'], {
       cwd,
       stdio: 'inherit'
     });
