+++ playground/custom-routing/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "custom-routing",
+  "version": "1.0.0",
+  "description": "",
+  "type": "module",
+  "private": "true",
+  "scripts": {
+    "dev": "abell --version && abell dev",
+    "generate": "abell generate"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "abell": "workspace:*"
+  }
+}
