+++ pnpm-workspace.yaml
@@ -1,5 +1,5 @@
 packages:
   - 'packages/abell'
   - 'packages/create-abell'
-  - 'packages/playground'
+  - 'playground/**'
   - 'packages/docs'
